replace mutex with atomic and separate tests with integration tag + minor fixes
