+++ .github/workflows/test.yaml
@@ -13,6 +13,6 @@ jobs:
         with:
           go-version: 1.17.x
       - name: Run coverage
-        run: go test -race -coverprofile=coverage.txt -covermode=atomic ./...
+        run: go test -race -coverprofile=coverage.txt -covermode=atomic --tags=integration ./...
       - name: Upload coverage to Codecov
         run: bash <(curl -s https://codecov.io/bash)
